Create 7. Final Project Enumeration Test
Also created an application that contains an enumeration (enum) that represents the days of the week.  Display a list of the days, then prompt the user for a day. Display business hours for the chosen day.  Create your own business hours stored in an array; however, every day of the week should have different hours.